requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
folium>=0.14.0
pyrosm>=0.6.0
shapely>=2.0.0
ortools>=9.7.0
requests>=2.31.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9